Add new light theme
changed the _apply_base_theme to apply AFTER palette is set, because you can't change the theme until setPalette is called. You won't notice this when there's only one theme, but when you try to switch between 2, it becomes apparent that the change only succeeds after you setPalette twice, which is due to it being called AFTER _apply_base_theme, instead of before.