Update application window title in setWindowTitle()
The application might not have a border to see the default title, but it is an application and visible on the taskbar. Rest assured that the system WILL add a title to your application, and your taskbar application will say "Python" instead of the proper window title, which is UGLY. We SHOULD be setting the applications title as well. There's no way to set this without ugly code, it belongs in the official module itself